move toolbar up one directory
