+++ packages/core/admin/admin/src/content-manager/components/BlocksEditor/tests/Toolbar.test.tsx
